\name{unlockEnvironment}
\alias{unlockEnvironment}
\title{Binding and Environment Unlocking}
\description{
  \code{unlockEnvironment()} unlocks an environment and possibly its bindings.
}
\usage{
unlockEnvironment(env, bindings = FALSE)
}
\arguments{
  \item{env}{an environment.}

  \item{bindings}{logical specifying whether bindings should be unlocked.}
}
\value{
  \code{NULL}, invisibly.
}
